Add more robust input checking in commands

+++ PrimaRoleReactions/role_reactions/src/commands.rs
@@ -14,13 +14,42 @@ use serenity::{
 #[commands(role_reactions, add_role_reaction, remove_role_reaction)]
 pub struct RoleReactions;
 
-fn read_role_reaction_info(message: &Message, mut args: Args) -> RoleReactionInfo {
-    RoleReactionInfo {
-        guild_id: *message.guild_id.unwrap().as_u64(),
-        channel_id: args.single::<u64>().unwrap(),
-        emoji_id: args.single::<u64>().unwrap(),
-        role_id: args.single::<u64>().unwrap(),
+async fn read_role_reaction_info(
+    ctx: &Context,
+    message: &Message,
+    mut args: Args,
+) -> Option<RoleReactionInfo> {
+    let guild_id = *message.guild_id.unwrap().as_u64();
+    let channel_id = args.single::<u64>();
+    let emoji_id = args.single::<u64>();
+    let role_id = args.single::<u64>();
+
+    let mut error_message: Option<&str> = None;
+    if let Err(_) = role_id {
+        error_message = Some("Failed to parse role ID.");
     }
+
+    if let Err(_) = emoji_id {
+        error_message = Some("Failed to parse emoji ID. Only non-Unicode emoji are supported.");
+    }
+
+    if let Err(_) = channel_id {
+        error_message = Some("Failed to parse channel ID.");
+    }
+
+    if let Some(error) = error_message {
+        println!("{}", error);
+        if let Err(why) = message.channel_id.say(&ctx.http, error).await {
+            println!("Error sending message: {:?}", why);
+        }
+    }
+
+    Some(RoleReactionInfo {
+        guild_id,
+        channel_id: channel_id.unwrap(),
+        emoji_id: emoji_id.unwrap(),
+        role_id: role_id.unwrap(),
+    })
 }
 
 #[command("rolereactions")]
@@ -62,18 +91,18 @@ async fn add_role_reaction(ctx: &Context, message: &Message, args: Args) -> Comm
         .get::<RoleReactionsDatabaseContainer>()
         .expect("Expected RoleReactionsDatabaseContainer in TypeMap");
 
-    let role_reaction = read_role_reaction_info(message, args);
-
-    match db.add_role_reaction(role_reaction).await {
-        Ok(_) => {}
-        Err(error) => {
-            println!("Failed to add role reaction: {:?}", error);
-            if let Err(why) = message
-                .channel_id
-                .say(&ctx.http, "Failed to add role reaction.")
-                .await
-            {
-                println!("Error sending message: {:?}", why);
+    if let Some(role_reaction) = read_role_reaction_info(ctx, message, args).await {
+        match db.add_role_reaction(role_reaction).await {
+            Ok(_) => {}
+            Err(error) => {
+                println!("Failed to add role reaction: {:?}", error);
+                if let Err(why) = message
+                    .channel_id
+                    .say(&ctx.http, "Failed to add role reaction.")
+                    .await
+                {
+                    println!("Error sending message: {:?}", why);
+                }
             }
         }
     }
@@ -88,18 +117,18 @@ async fn remove_role_reaction(ctx: &Context, message: &Message, args: Args) -> C
         .get::<RoleReactionsDatabaseContainer>()
         .expect("Expected RoleReactionsDatabaseContainer in TypeMap");
 
-    let role_reaction = read_role_reaction_info(message, args);
-
-    match db.remove_role_reaction(role_reaction).await {
-        Ok(_) => {}
-        Err(error) => {
-            println!("Failed to remove role reaction: {:?}", error);
-            if let Err(why) = message
-                .channel_id
-                .say(&ctx.http, "Failed to remove role reaction.")
-                .await
-            {
-                println!("Error sending message: {:?}", why);
+    if let Some(role_reaction) = read_role_reaction_info(ctx, message, args).await {
+        match db.remove_role_reaction(role_reaction).await {
+            Ok(_) => {}
+            Err(error) => {
+                println!("Failed to remove role reaction: {:?}", error);
+                if let Err(why) = message
+                    .channel_id
+                    .say(&ctx.http, "Failed to remove role reaction.")
+                    .await
+                {
+                    println!("Error sending message: {:?}", why);
+                }
             }
         }
     }
